requirements.txt
aiogram==3.1.0
aiohttp==3.8.5
craiyon.py==1.2.0
